backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
